third_party/blink/web_tests/external/wpt/custom-elements/HTMLElement-constructor-expected.txt

This is a testharness.js-based test.
[FAIL] HTMLElement constructor must only get .prototype once, calling proxy constructor directly
  assert_equals: Should have gotten .prototype once expected 1 but got 2
[FAIL] HTMLElement constructor must only get .prototype once, calling proxy constructor via Reflect
  assert_equals: Should have gotten .prototype once expected 1 but got 2
[FAIL] HTMLElement constructor must only get .prototype once, calling proxy constructor via Reflect with no inheritance
  assert_equals: Should have gotten .prototype once expected 1 but got 2
[FAIL] HTMLElement constructor must not get .prototype until it finishes its registration  sanity checks, calling via Reflect
  assert_equals: Should never have gotten .prototype expected 0 but got 1
Harness: the test ran to completion.
